file rename and linting
