+++ ui/src/app/pages/analysis/interactive-notebook.tsx
@@ -2,12 +2,12 @@ import * as Cookies from 'js-cookie';
 import * as fp from 'lodash/fp';
 import * as React from 'react';
 
+import {IconButton} from 'app/components/buttons';
 import {ClrIcon} from 'app/components/icons';
+import {PlaygroundIcon} from 'app/components/icons';
 import {TooltipTrigger} from 'app/components/popups';
 import {SpinnerOverlay} from 'app/components/spinners';
 import {EditComponentReact} from 'app/icons/edit';
-import {PlaygroundIcon} from 'app/components/icons';
-import {IconButton} from 'app/components/buttons'
 import {ConfirmPlaygroundModeModal} from 'app/pages/analysis/confirm-playground-mode-modal';
 import {NotebookInUseModal} from 'app/pages/analysis/notebook-in-use-modal';
 import {workspacesApi} from 'app/services/swagger-fetch-clients';
